refactor: split lib/constants/ui/index.ts into themes and gameSteps subfiles (#498)
Extract GAME_BACKGROUNDS and START_BUTTON_COLORS into uiData/themes.ts
and all 21 *_GAME_STEPS arrays into uiData/gameSteps.ts. Reduce index.ts
to a 2-line barrel re-exporting both subfiles. No consumer import paths
changed.
