+++ gamesformykids/lib/constants/ui/uiData/themes.ts
@@ -1,0 +1,15 @@
+export const GAME_BACKGROUNDS = {
+  LETTERS: "linear-gradient(135deg, #fed7aa 0%, #fdba74 25%, #fb923c 50%, #f97316 75%, #ea580c 100%)",
+  NUMBERS: "linear-gradient(135deg, #e0e7ff 0%, #c7d2fe 25%, #a5b4fc 50%, #818cf8 75%, #6366f1 100%)",
+  SHAPES: "linear-gradient(135deg, #d4f1d4 0%, #a8e6a8 25%, #7dd87d 50%, #52c952 75%, #26b926 100%)",
+  COLORS: "linear-gradient(135deg, #ffecd2 0%, #fcb69f 25%, #a8e6cf 50%, #dcedc1 75%, #ffd3e1 100%)",
+  MEMORY: "linear-gradient(135deg, #fce7f3 0%, #e879f9 25%, #a855f7 50%, #7c3aed 75%, #5b21b6 100%)",
+};
+
+export const START_BUTTON_COLORS = {
+  LETTERS: { from: "yellow", to: "orange" },
+  NUMBERS: { from: "indigo", to: "purple" },
+  SHAPES: { from: "blue", to: "green" },
+  COLORS: { from: "teal", to: "cyan" },
+  MEMORY: { from: "fuchsia", to: "pink" },
+};
